Long 字段调整

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/domain/Action.java
@@ -108,7 +108,7 @@ public class Action extends EntityMP implements Serializable {
     @TableField(value = "objectid")
     @JSONField(name = "objectid")
     @JsonProperty("objectid")
-    private Integer objectid;
+    private Long objectid;
     /**
      * 操作者
      */
@@ -143,7 +143,7 @@ public class Action extends EntityMP implements Serializable {
     @TableField(exist = false)
     @JSONField(name = "isactorss")
     @JsonProperty("isactorss")
-    private Integer isactorss;
+    private Long isactorss;
     /**
      * 显示日期
      */
@@ -213,7 +213,7 @@ public class Action extends EntityMP implements Serializable {
     /**
      * 设置 [对象ID]
      */
-    public void setObjectid(Integer objectid){
+    public void setObjectid(Long objectid){
         this.objectid = objectid ;
         this.modify("objectid",objectid);
     }
